tests: Synthetic tests moved up and out of test infra
The disruption suite needs to skip many of the synthetics because
the invariants aren't expected to be maintained. Hoist those use
cases up and out of the test/ginkgo package and rename them to
"invariants", using the same mechanism we use to inject the external
image checker.
